+++ example/wallet/lib/dependencies/web3wallet_service.dart
@@ -10,6 +10,8 @@ import 'package:walletconnect_flutter_v2_wallet/dependencies/deep_link_handler.d
 import 'package:walletconnect_flutter_v2_wallet/dependencies/i_web3wallet_service.dart';
 import 'package:walletconnect_flutter_v2_wallet/dependencies/key_service/chain_key.dart';
 import 'package:walletconnect_flutter_v2_wallet/dependencies/key_service/i_key_service.dart';
+import 'package:walletconnect_flutter_v2_wallet/models/chain_metadata.dart';
+import 'package:walletconnect_flutter_v2_wallet/utils/constants.dart';
 import 'package:walletconnect_flutter_v2_wallet/utils/dart_defines.dart';
 import 'package:walletconnect_flutter_v2_wallet/widgets/wc_connection_request/wc_auth_request_model.dart';
 import 'package:walletconnect_flutter_v2_wallet/widgets/wc_connection_request/wc_connection_request_widget.dart';
@@ -34,7 +36,7 @@ class Web3WalletService extends IWeb3WalletService {
   ValueNotifier<List<StoredCacao>> auth = ValueNotifier<List<StoredCacao>>([]);
 
   @override
-  void create() {
+  void create() async {
     // Create the web3wallet
     _web3Wallet = Web3Wallet(
       core: Core(
@@ -42,11 +44,11 @@ class Web3WalletService extends IWeb3WalletService {
         logLevel: LogLevel.error,
       ),
       metadata: const PairingMetadata(
-        name: 'Web3Wallet Flutter Example',
-        description: 'Web3Wallet Flutter Example',
+        name: 'Sample Wallet Flutter',
+        description: 'WalletConnect\'s sample wallet with Flutter',
         url: 'https://walletconnect.com/',
         icons: [
-          'https://raw.githubusercontent.com/WalletConnect/WalletConnectFlutterV2/master/example/wallet/AppIcon.png'
+          'https://docs.walletconnect.com/assets/images/web3walletLogo-54d3b546146931ceaf47a3500868a73a.png'
         ],
         redirect: Redirect(
           native: 'wcflutterwallet://',
@@ -56,18 +58,22 @@ class Web3WalletService extends IWeb3WalletService {
     );
 
     // Setup our accounts
-    List<ChainKey> chainKeys = GetIt.I<IKeyService>().getKeys();
+    List<ChainKey> chainKeys = await GetIt.I<IKeyService>().setKeys();
+    if (chainKeys.isEmpty) {
+      await GetIt.I<IKeyService>().createWallet();
+      chainKeys = await GetIt.I<IKeyService>().setKeys();
+    }
     for (final chainKey in chainKeys) {
       for (final chainId in chainKey.chains) {
         if (chainId.startsWith('kadena')) {
           _web3Wallet!.registerAccount(
             chainId: chainId,
-            accountAddress: 'k**${chainKey.publicKey}',
+            accountAddress: 'k**${chainKey.address}',
           );
         } else {
           _web3Wallet!.registerAccount(
             chainId: chainId,
-            accountAddress: chainKey.publicKey,
+            accountAddress: chainKey.address,
           );
         }
       }
@@ -123,18 +129,75 @@ class Web3WalletService extends IWeb3WalletService {
     debugPrint('[$runtimeType] _onRelayClientError ${args?.error}');
   }
 
-  void _onSessionProposalError(SessionProposalErrorEvent? args) {
+  void _onSessionProposalError(SessionProposalErrorEvent? args) async {
     debugPrint('[$runtimeType] _onSessionProposalError $args');
+    if (args != null) {
+      String errorMessage = args.error.message;
+      if (args.error.code == 5100) {
+        errorMessage =
+            errorMessage.replaceFirst('Requested:', '\n\nRequested:');
+        errorMessage =
+            errorMessage.replaceFirst('Supported:', '\n\nSupported:');
+      }
+      GetIt.I<IBottomSheetService>().queueBottomSheet(
+        widget: Container(
+          color: Colors.white,
+          width: double.infinity,
+          padding: const EdgeInsets.all(20.0),
+          child: Column(
+            children: [
+              Icon(
+                Icons.error_outline_sharp,
+                color: Colors.red[100],
+                size: 80.0,
+              ),
+              Text(
+                'Error',
+                style: StyleConstants.subtitleText.copyWith(
+                  color: Colors.black,
+                  fontSize: 18.0,
+                ),
+              ),
+              Text(errorMessage),
+            ],
+          ),
+        ),
+      );
+    }
+  }
+
+  Map<String, Namespace> _generateNamespaces(
+    Map<String, Namespace>? approvedNamespaces,
+    ChainType chainType,
+  ) {
+    //
+    final constructedNS = Map<String, Namespace>.from(approvedNamespaces ?? {});
+    constructedNS[chainType.name] = constructedNS[chainType.name]!.copyWith(
+      methods: [
+        'personal_sign',
+        ...constructedNS[chainType.name]!.methods,
+      ],
+    );
+    return constructedNS;
   }
 
   void _onSessionProposal(SessionProposalEvent? args) async {
     if (args != null) {
+      // generatedNamespaces is constructed based on registered methods handlers
+      // so if you want to handle requests using onSessionRequest event then you would need to manually add that method in the approved namespaces
+      final approvedNS = _generateNamespaces(
+        args.params.generatedNamespaces!,
+        ChainType.eip155,
+      );
+      final proposalData = args.params.copyWith(
+        generatedNamespaces: approvedNS,
+      );
       final approved = await _bottomSheetHandler.queueBottomSheet(
         widget: WCRequestWidget(
           child: WCConnectionRequestWidget(
             wallet: _web3Wallet!,
             sessionProposal: WCSessionRequestModel(
-              request: args.params,
+              request: proposalData,
               verifyContext: args.verifyContext,
             ),
           ),
@@ -144,17 +207,28 @@ class Web3WalletService extends IWeb3WalletService {
       if (approved == true) {
         _web3Wallet!.approveSession(
           id: args.id,
-          namespaces: args.params.generatedNamespaces!,
+          namespaces: approvedNS,
         );
+        final scheme = args.params.proposer.metadata.redirect?.native ?? '';
+        DeepLinkHandler.goTo(scheme, delay: 300);
       } else {
         _web3Wallet!.rejectSession(
           id: args.id,
           reason: Errors.getSdkError(Errors.USER_REJECTED),
         );
-      }
+        _web3Wallet!.core.pairing.disconnect(
+          topic: args.params.pairingTopic,
+        );
 
-      final scheme = args.params.proposer.metadata.redirect?.native ?? '';
-      DeepLinkHandler.goTo(scheme, delay: 300);
+        final scheme = args.params.proposer.metadata.redirect?.native ?? '';
+        DeepLinkHandler.goTo(
+          scheme,
+          delay: 300,
+          modalTitle: 'Error',
+          modalMessage: 'User rejected',
+          success: false,
+        );
+      }
     }
   }
 
@@ -168,11 +242,9 @@ class Web3WalletService extends IWeb3WalletService {
 
   Future<void> _onAuthRequest(AuthRequest? args) async {
     if (args != null) {
-      List<ChainKey> chainKeys = GetIt.I<IKeyService>().getKeysForChain(
-        'eip155:1',
-      );
+      final chainKeys = GetIt.I<IKeyService>().getKeysForChain('eip155:1');
       // Create the message to be signed
-      final String iss = 'did:pkh:eip155:1:${chainKeys.first.publicKey}';
+      final iss = 'did:pkh:eip155:1:${chainKeys.first.address}';
 
       final bool? auth = await _bottomSheetHandler.queueBottomSheet(
         widget: WCRequestWidget(